refactor: change place definition of embeddings
