+++ tests/unit/types/arrays/test_memmap.py
@@ -467,3 +467,38 @@ def test_split(memmap_for_split):
     assert len(rv['c']) == 2
     assert len(rv['b']) == 1
     assert len(rv['a']) == 2
+
+
+def test_dam_embeddings(tmpdir):
+    dam = DocumentArrayMemmap(tmpdir)
+    dam.extend(Document(embedding=np.array([1, 2, 3, 4])) for _ in range(100))
+    np.testing.assert_almost_equal(dam.get_attributes('embedding'), dam.embeddings)
+
+
+def test_dam_get_embeddings(tmpdir):
+    da = DocumentArrayMemmap(tmpdir)
+    da.extend(Document(embedding=np.array([1, 2, 3, 4])) for _ in range(100))
+    np.testing.assert_almost_equal(
+        da.get_attributes('embedding')[10:20], da._get_embeddings(slice(10, 20))
+    )
+
+
+def test_embeddings_setter_dam(tmpdir):
+    emb = np.random.random((100, 128))
+    dam = DocumentArrayMemmap(tmpdir)
+    dam.extend([Document() for _ in range(100)])
+    dam.embeddings = emb
+    np.testing.assert_almost_equal(dam.embeddings, emb)
+
+    for x, doc in zip(emb, dam):
+        np.testing.assert_almost_equal(x, doc.embedding)
+
+
+def test_embeddings_getter_dam(tmpdir):
+    emb = np.random.random((100, 128))
+    dam = DocumentArrayMemmap(tmpdir)
+    dam.extend([Document(embedding=x) for x in emb])
+
+    np.testing.assert_almost_equal(dam.embeddings, emb)
+    for x, doc in zip(emb, dam):
+        np.testing.assert_almost_equal(x, doc.embedding)
