+++ tests/unit/types/arrays/test_neural_ops.py
@@ -9,7 +9,6 @@ import pytest
 from jina import Document, DocumentArray
 from jina.types.arrays.memmap import DocumentArrayMemmap
 from jina.math.dimensionality_reduction import PCA
-from tests import random_docs
 
 
 @pytest.fixture()
@@ -399,69 +398,3 @@ def test_match_inclusive_dam(tmpdir):
     assert len(da2) == 3
     traversed = dam.traverse_flat(traversal_paths=['m', 'mm', 'mmm'])
     assert len(list(traversed)) == 9
-
-
-def test_da_get_embeddings():
-    da = DocumentArray(random_docs(100))
-    np.testing.assert_almost_equal(da.get_attributes('embedding'), da.embeddings)
-
-
-def test_dam_embeddings(tmpdir):
-    dam = DocumentArrayMemmap(tmpdir)
-    dam.extend(Document(embedding=np.array([1, 2, 3, 4])) for _ in range(100))
-    np.testing.assert_almost_equal(dam.get_attributes('embedding'), dam.embeddings)
-
-
-def test_da_get_embeddings():
-    da = DocumentArray(random_docs(100))
-    np.testing.assert_almost_equal(
-        da.get_attributes('embedding')[10:20], da._get_embeddings(slice(10, 20))
-    )
-
-
-def test_dam_get_embeddings(tmpdir):
-    da = DocumentArrayMemmap(tmpdir)
-    da.extend(Document(embedding=np.array([1, 2, 3, 4])) for _ in range(100))
-    np.testing.assert_almost_equal(
-        da.get_attributes('embedding')[10:20], da._get_embeddings(slice(10, 20))
-    )
-
-
-def test_embeddings_setter_da():
-    emb = np.random.random((100, 128))
-    da = DocumentArray([Document() for _ in range(100)])
-    da.embeddings = emb
-    np.testing.assert_almost_equal(da.embeddings, emb)
-
-    for x, doc in zip(emb, da):
-        np.testing.assert_almost_equal(x, doc.embedding)
-
-
-def test_embeddings_setter_dam(tmpdir):
-    emb = np.random.random((100, 128))
-    dam = DocumentArrayMemmap(tmpdir)
-    dam.extend([Document() for _ in range(100)])
-    dam.embeddings = emb
-    np.testing.assert_almost_equal(dam.embeddings, emb)
-
-    for x, doc in zip(emb, dam):
-        np.testing.assert_almost_equal(x, doc.embedding)
-
-
-def test_embeddings_getter_da():
-    emb = np.random.random((100, 128))
-    da = DocumentArray([Document(embedding=x) for x in emb])
-    np.testing.assert_almost_equal(da.embeddings, emb)
-
-    for x, doc in zip(emb, da):
-        np.testing.assert_almost_equal(x, doc.embedding)
-
-
-def test_embeddings_getter_dam(tmpdir):
-    emb = np.random.random((100, 128))
-    dam = DocumentArrayMemmap(tmpdir)
-    dam.extend([Document(embedding=x) for x in emb])
-
-    np.testing.assert_almost_equal(dam.embeddings, emb)
-    for x, doc in zip(emb, dam):
-        np.testing.assert_almost_equal(x, doc.embedding)
